requirements.txt
# Flask web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0

# Environment variables
python-dotenv==1.0.0

# Basic utilities
requests==2.31.0

# Google Gemini for typo checking (REQUIRED)
google-generativeai==0.3.2

# ML stack for sentiment/entity analysis
numpy>=1.26.0
torch>=2.2.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
transformers==4.35.2
huggingface_hub==0.16.4
